# backend/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Access API",
    "version": "1.0.0"
  },
  "paths": {
    "/": {
      "get": {
        "description": "Returns message showing app is up",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "hello world"
          }
        }
      }
    },
      "/register": {
        "post": {
          "tags": ["User"],
          "description": "Register a new user",
          "produces": ["application/json"],
          "parameters": [
            {
              "name": "username",
              "in": "formData",
              "description": "Username of the user",
              "required": true,
              "type": "string"
            },
            {
              "name": "email",
              "in": "formData",
              "description": "Email of the user",
              "required": true,
              "type": "string"
            },
            {
              "name": "password",
              "in": "formData",
              "description": "Password of the user",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "201": {
              "description": "User registered successfully"
            },
            "400": {
              "description": "Invalid input"
            },
            "409": {
              "description": "User already exists"
            }
          }
        }
      },
      "/login": {
        "post": {
        "tags": ["User"],
          "description": "Login a user",
          "produces": ["application/json"],
          "parameters": [
            {
              "name": "email",
              "in": "formData",
              "description": "Email of the user",
              "required": true,
              "type": "string"
            },
            {
              "name": "password",
              "in": "formData",
              "description": "Password of the user",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "User logged in successfully"
            },
            "400": {
              "description": "Invalid input"
            },
            "404": {
              "description": "User not found"
            }
          }
        }
      },
      "/mapbox/suggest": {
        "get": {
        "tags": ["MapBox"],
          "description": "Get suggestions for a location",
          "produces": ["application/json"],
          "parameters": [
            {
              "name": "query",
              "in": "query",
              "description": "Query string for location",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Location suggestions"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/mapbox/retrieve": {
        "get": {
        "tags": ["MapBox"],
          "description": "Retrieve location details",
          "produces": ["application/json"],
          "parameters": [
            {
              "name": "mapbox_id",
              "in": "query",
              "description": "Place MapBox ID of the location",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Location details"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/mapbox/trip": {
        "get": {
        "tags": ["MapBox"],
          "description": "Get trip itinerary",
          "produces": ["application/json"],
          "parameters": [
            {
              "name": "depart_lat",
              "in": "query",
              "description": "Departure latitude",
              "required": true,
              "type": "number"
            },
            {
              "name": "depart_lon",
              "in": "query",
              "description": "Departure longitude",
              "required": true,
              "type": "number"
            },
            {
              "name": "arrive_lat",
              "in": "query",
              "description": "Arrival latitude",
              "required": true,
              "type": "number"
            },
            {
              "name": "arrive_lon",
              "in": "query",
              "description": "Arrival longitude",
              "required": true,
              "type": "number"
            }
          ],
          "responses": {
            "200": {
              "description": "Trip itinerary"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      }
  }
}